ai-code-reviewer/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
google-genai==0.3.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
pydantic==2.9.2
